refactor(style): separa estilos de layout das páginas de autenticação do layout de inscrição
