+++ src/app/(auth)/register/additional-info/page.jsx
@@ -10,7 +10,7 @@ import { createAdditionalInfoSchema } from '@/schemas/register.schema';
 import { useRegister } from '@/contexts/RegisterContext';
 import { useFormStage } from '@/hooks/useFormStage';
 import Button from '@/components/Button';
-import styles from '@/app/(auth)/register/layout.module.css';
+import styles from '@/app/(auth)/layout.module.css';
 import { useAuthLayout } from '@/contexts/AuthLayoutContext';
 
 export default function AdditionalInfoStep() {
